requirements.txt
aiogram==3.24.0
requests==2.32.5
python-dotenv==1.2.1
